updated list with dogs form
