+++ webapp/src/commoncomponents/grids/shelterDogGrid.test.tsx
@@ -7,8 +7,8 @@ import { render, unmountComponentAtNode } from "react-dom";
 import { act, createRenderer } from "react-dom/test-utils";
 import { isNull } from "lodash";
 import { Button } from "@material-ui/core";
-import ImageGrid from "./imageGrid";
-import { testDogList } from "../dog/dogTesting";
+import ShelterDogGrid from "./shelterDogGrid";
+import { otherDog, sheltertestDog, testDogList } from "../../dog/dogTesting";
 import { Provider } from "react-redux";
 import {
   Redirect,
@@ -17,7 +17,7 @@ import {
   Switch,
   useHistory,
 } from "react-router-dom";
-import { store } from "../app/store";
+import { store } from "../../app/store";
 
 let container: HTMLDivElement | null = null;
 beforeEach(() => {
@@ -41,7 +41,7 @@ it("Rendered list of dog cards", () => {
       <Provider store={store}>
         <Router>
           <Route path="/">
-            <ImageGrid dogs={testDogList} id={-1} cookies={{}} path={""} />
+            <ShelterDogGrid dogs={testDogList} id={-1} cookies={{}} path={""} />
           </Route>
         </Router>
       </Provider>,
@@ -55,4 +55,5 @@ it("Rendered list of dog cards", () => {
   ).toBeGreaterThanOrEqual(testDogList.length);
   //check if dog named alex was rendered
   expect(container?.getElementsByClassName("Alex").length).toEqual(1);
+  expect(container?.getElementsByClassName("Alexa").length).toEqual(1);
 });
