+++ webapp/src/dog/registerDog/registerDog.tsx
@@ -17,7 +17,7 @@ import { createStyles, makeStyles, Theme } from "@material-ui/core/styles";
 import { DatePicker, MuiPickersUtilsProvider } from "@material-ui/pickers";
 import InputLabel from "@material-ui/core/InputLabel";
 import DateFnsUtils from "@date-io/date-fns";
-import ImageUpload from "./ImageUpload";
+import ImageUpload from "../../commonComponents/ImageUpload";
 import {
   ColorTypes,
   HairTypes,
