+++ webapp/src/shelter/adoptDogList/listWithAdoptDogs.tsx
@@ -27,7 +27,7 @@ import Toolbar from "@material-ui/core/Toolbar";
 import CssBaseline from "@material-ui/core/CssBaseline";
 import { useSelector } from "react-redux";
 import config from "../../config/config";
-import ImageGrid from "../../commonComponents/imageGrid";
+import ImageGrid from "../../commonComponents/grids/imageGrid";
 import InfiniteScroll from "react-infinite-scroll-component";
 import SendIcon from "@material-ui/icons/Send";
 import { clearDogList, logoutThunk } from "../../app/actions";
@@ -46,8 +46,8 @@ import ContactInfo from "../../contactInfo/contactInformation";
 import { IContactInfo } from "../../contactInfo/contactInfoInterface";
 import { IShelterDogWithPicture } from "../../dog/dogInterfaces";
 import ShelterListWithDogs from "../shelterDogsList/shelterListWithDogs";
-import SheltersGrid from "../../commonComponents/sheltersGrid";
-import ShelterDogGrid from "../../commonComponents/shelterDogGrid";
+import SheltersGrid from "../../commonComponents/grids/sheltersGrid";
+import ShelterDogGrid from "../../commonComponents/grids/shelterDogGrid";
 import { State } from "../../app/stateInterfaces";
 import { IShelter } from "../shelterInterfaces";
 const SidebarTrigger = getSidebarTrigger(styled);
