+++ webapp/src/shelter/shelterDogsList/shelterListWithDogs.tsx
@@ -30,7 +30,7 @@ import { clearDogList, logoutThunk } from "../../app/actions";
 import LoadingPopup from "../../utilityComponents/LoadingPopup";
 import Footer from "../../utilityComponents/Footer";
 import { IFilterSort } from "../../dog/dogsList/filterInterface";
-import ShelterDogGrid from "../../commonComponents/shelterDogGrid";
+import ShelterDogGrid from "../../commonComponents/grids/shelterDogGrid";
 import { State } from "../../app/stateInterfaces";
 
 const Content = getContent(styled);
